Moved Sqlite DbDescribe cests to individual method cests.
