+++ tests/integration/Db/Adapter/Pdo/Sqlite/DescribeReferencesCest.php
@@ -13,12 +13,21 @@ declare(strict_types=1);
 namespace Phalcon\Test\Integration\Db\Adapter\Pdo\Sqlite;
 
 use IntegrationTester;
+use Phalcon\Db\Reference;
+use Phalcon\Test\Fixtures\Traits\DiTrait;
 
-/**
- * Class DescribeReferencesCest
- */
 class DescribeReferencesCest
 {
+    use DiTrait;
+
+    public function _before(IntegrationTester $I)
+    {
+        $this->newDi();
+        $this->setDiModelsManager();
+        $this->setDiModelsMetadata();
+        $this->setDiSqlite();
+    }
+
     /**
      * Tests Phalcon\Db\Adapter\Pdo\Sqlite :: describeReferences()
      *
@@ -27,7 +36,32 @@ class DescribeReferencesCest
      */
     public function dbAdapterPdoSqliteDescribeReferences(IntegrationTester $I)
     {
-        $I->wantToTest('Db\Adapter\Pdo\Sqlite - describeReferences()');
-        $I->skipTest('Need implementation');
+        $connection = $this->getService('db');
+
+        $expectedReferences = [
+            'foreign_key_0' => Reference::__set_state(
+                [
+                    'referenceName'     => 'foreign_key_0',
+                    'referencedTable'   => 'parts',
+                    'columns'           => ['parts_id'],
+                    'referencedColumns' => ['id'],
+                    'referencedSchema'  => null,
+                ]
+            ),
+            'foreign_key_1' => Reference::__set_state(
+                [
+                    'referenceName'     => 'foreign_key_1',
+                    'referencedTable'   => 'robots',
+                    'columns'           => ['robots_id'],
+                    'referencedColumns' => ['id'],
+                    'referencedSchema'  => null,
+                ]
+            ),
+        ];
+
+        $I->assertEquals(
+            $expectedReferences,
+            $connection->describeReferences('robots_parts')
+        );
     }
 }
