+++ tests/integration/Db/Adapter/Pdo/Sqlite/TableExistsCest.php
@@ -13,12 +13,20 @@ declare(strict_types=1);
 namespace Phalcon\Test\Integration\Db\Adapter\Pdo\Sqlite;
 
 use IntegrationTester;
+use Phalcon\Test\Fixtures\Traits\DiTrait;
 
-/**
- * Class TableExistsCest
- */
 class TableExistsCest
 {
+    use DiTrait;
+
+    public function _before(IntegrationTester $I)
+    {
+        $this->newDi();
+        $this->setDiModelsManager();
+        $this->setDiModelsMetadata();
+        $this->setDiSqlite();
+    }
+
     /**
      * Tests Phalcon\Db\Adapter\Pdo\Sqlite :: tableExists()
      *
@@ -28,6 +36,23 @@ class TableExistsCest
     public function dbAdapterPdoSqliteTableExists(IntegrationTester $I)
     {
         $I->wantToTest('Db\Adapter\Pdo\Sqlite - tableExists()');
-        $I->skipTest('Need implementation');
+
+        $connection = $this->getService('db');
+
+        $I->assertTrue(
+            $connection->tableExists('personas')
+        );
+
+        $I->assertFalse(
+            $connection->tableExists('noexist')
+        );
+
+        $I->assertTrue(
+            $connection->tableExists('personas', 'public')
+        );
+
+        $I->assertTrue(
+            $connection->tableExists('personas', 'test')
+        );
     }
 }
